unit tests

+++ CurrencyConverter/src/core/CurrencyConverter.Core/Models/CurrencyConverterResponse.cs
@@ -2,7 +2,7 @@ namespace CurrencyConverter.Core.Models
 {
     public class CurrencyConverterResponse
     {
-        public float Result { get; set; }
+        public double Result { get; set; }
         public ResponseResult ResponseResult { get; set; }
     }
 }
